Updated azure backend

+++ skyway/cloud/azure.py
@@ -74,24 +74,34 @@ class AZURE(Cloud):
         Get a list of all existed instances
         
         Return: a list of multiple turple. Each turple has four elements:
-                (1) instance name (2) state (3) type (4) identifier
+                (1) instance name (2) state (3) type (4) running time
+                node.id is not particular useful
         """
         nodes = []
         current_time = datetime.now(timezone.utc)
         for node in self.driver.list_nodes():
-
+            
             # Get the creation time of the instance
             creation_time_str = node.extra.get('properties')['timeCreated']  # Azure
             if creation_time_str:
                 # Convert the creation time from string to datetime object
+                # Azure returns 7-digit after '.' for seconds, so need to truncate the last digit 
+                # to cast into %Y-%m-%dT%H:%M:%S.%f%z format
+                idx = creation_time_str.find('+')
+                creation_time_str = creation_time_str[:idx-1] + creation_time_str[idx:]
+                
                 creation_time = datetime.strptime(creation_time_str, '%Y-%m-%dT%H:%M:%S.%f%z')
                 
+                # get the node type
+                node_type = node.extra.get('properties')['hardwareProfile']['vmSize']
+
                 # Calculate the running time
-                running_time = current_time - creation_time               
-                nodes.append([node.name, node.state, node.size, node.id, node.public_ips[0], running_time])
+                running_time = current_time - creation_time
+                
+                nodes.append([node.name, node.state, node_type, running_time])
 
         if verbose == True:
-            print(tabulate(nodes, headers=['Name', 'Status', 'Type', 'Instance ID', 'Host', 'Running Time']))
+            print(tabulate(nodes, headers=['Name', 'Status', 'Type', 'Running Time']))
             print("")
 
 
@@ -111,13 +121,16 @@ class AZURE(Cloud):
                 raise ValueError(f"Location '{location_name}' not found.")
 
             # authentication with public key on this machine per-user (id_rsa_azure.pub)
+            # need to read in from ~/.ssh/id_rsa_azure.pub from the account .yaml file
             auth = NodeAuthSSHKey('ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQDXzS0IuhNV3EyNkUXRV0yML0Fi+r+7qkqQqC7Ahe239ct3wwj1uwFnHo6UxF7zH7i33rtP/0YX5aV0l1fxp/7S3pF7Y0ZQDYryW15DxDxxjHkbNEssZjQ0XYWtf6g6VD5v8UWZAFvctqytbE2f39xDRimvifQMh8ogG45zRquKfuZ1vmtT/ls/7iW4cRtVjXNhN+lEmggZ+183akqDE0OJ01SM0aVNULxo/CB/nZFibgZ6YgQ8Ak/h5d3cHRH2YHQE6Szqi9+jYn/+99xLzQTfu6fF4uV2xw7BJ+O6UKvvJhYQMS/LUV14xmWfwJxJX/4lUh3Yc58kWZp4GSTpdyMByU/ejrvsrDkbzmjwu/TgSrAADfMxBMnVHkLhg9hhCmYDAtlY79OxMPt5WQtIZcZJbBJNW5d6fuOM6dEvw7p3Qu/QNhgEbYXYnlW+izPVhcuqe2YvmnnLKYERPSQS4VAsZhboRilmCotfJXxSC7Uf4oDbRdBxnZRvwC6vssj3tIiBY9sHXSFZyOH1d0MhmrQzwt09L0GrMgdthVPdhWW/V19pvhFnV8UNpanJXy09IiuCrJsKYz7k4YfPiCJppPU9xXMcYyZ9QOLKDmPZbUpEySZAPf77AaeUnp6xcDb9zF7+iOx4xjw3ZzMuCQfWhefu7kLOPHml+OukupmPCP6jyQ== ndtrung@fedora')
+            # resource group is already created on the subscription
             resource_group = "rg_skyway"
 
             # Initialize Azure management clients
-            resource_client = ResourceManagementClient(self.credentials, self.subscription_id)
-            network_client = NetworkManagementClient(self.credentials, self.subscription_id)
-            compute_client = ComputeManagementClient(self.credentials, self.subscription_id)
+            subscription_id = self.account['subscription_id']
+            resource_client = ResourceManagementClient(self.credentials, subscription_id)
+            network_client = NetworkManagementClient(self.credentials, subscription_id)
+            compute_client = ComputeManagementClient(self.credentials, subscription_id)
 
             
             sizes = self.driver.list_sizes(location=location)
@@ -155,7 +168,7 @@ class AZURE(Cloud):
                                                         location=location)
             ip_config = {
                 "name": "ipconfig1",
-                "subnet": {"id": f"/subscriptions/{self.subscription_id}/resourceGroups/{resource_group}/providers/Microsoft.Network/virtualNetworks/{vnet_name}/subnets/{subnet_name}"},
+                "subnet": {"id": f"/subscriptions/{subscription_id}/resourceGroups/{resource_group}/providers/Microsoft.Network/virtualNetworks/{vnet_name}/subnets/{subnet_name}"},
                 'public_ip': public_ip
             }
         
@@ -163,7 +176,9 @@ class AZURE(Cloud):
                 "location": location_name,
                 "ip_configurations": [ip_config]
             }
-            network_interface = network_client.network_interfaces.begin_create_or_update(resource_group, nic_name, nic_params).result()
+            network_interface = network_client.network_interfaces.begin_create_or_update(resource_group,
+                                                                                         nic_name,
+                                                                                         nic_params).result()
 
             # Step 5: Create the instance          
             try:
@@ -178,23 +193,30 @@ class AZURE(Cloud):
             except Exception as ex:
                 logging.info("Failed to create %s. Reason: %s" % (node_name, str(ex)))
 
-            nodes[node_name] = [node_name, node.public_ips[0]]
+            nodes[node_name] = [node_name]
 
         return nodes
 
     def connect_node(self, node_name):
         pass
 
-    def destroy_node(self, node_names):
+    def destroy_nodes(self, node_names):
+        '''
+        Destroy all the nodes
+        node_names = list of node names as strings
+        '''
         if isinstance(node_names, str): node_names = [node_names]
-        
+
+        nodes = self.driver.list_nodes()
         for name in node_names:
-            try:
-                node = self.driver.ex_get_node(name)
+            #node = self.driver.ex_get_node(name)
+            node = next((nd for nd in nodes if nd.name == name), None)
+            if node is None:
+                raise ValueError(f"Node {name} not found.")
+                
+            response = input(f"Do you want to destroy {node.name}? (y/n) ")
+            if response == 'y':
                 self.driver.destroy_node(node)
-            except:
-                continue
-        return
         
     def check_valid_user(self, user_name, verbose=False):
         if user_name not in self.users:
@@ -230,17 +252,97 @@ class AZURE(Cloud):
         print("")
 
     def get_running_nodes(self, verbose=False):
-        pass
+        """Member function: list_nodes
+        Get a list of all existed instances
+        
+        Return: a list of multiple turple. Each turple has four elements:
+                (1) instance name (2) state (3) type (4) identifier
+        """
+        nodes = []
+        current_time = datetime.now(timezone.utc)
+        for node in self.driver.list_nodes():
+            if node.state == "running":
+                # Get the creation time of the instance
+                creation_time_str = node.extra.get('properties')['timeCreated']  # Azure
+                if creation_time_str:
+                    # Convert the creation time from string to datetime object
+                    # Azure returns 7-digit after '.' for seconds, so need to truncate the last digit 
+                    # to cast into %Y-%m-%dT%H:%M:%S.%f%z format
+                    idx = creation_time_str.find('+')
+                    creation_time_str = creation_time_str[:idx-1] + creation_time_str[idx:]
 
-    def get_unit_price(self, instance):
+                    creation_time = datetime.strptime(creation_time_str, '%Y-%m-%dT%H:%M:%S.%f%z')
+                    
+                    # Calculate the running time
+                    running_time = current_time - creation_time
+
+                    # get the node type
+                    node_type = node.extra.get('properties')['hardwareProfile']['vmSize']
+
+                    nodes.append([node.name, node.state, node_type, running_time])
+
+        if verbose == True:
+            print(tabulate(nodes, headers=['Name', 'Status', 'Type', 'Instance ID', 'Running Time']))
+            print("")
+
+    def get_unit_price(self, node):
         """
         Get the per-hour price of an instance depending on its instance_type (e.g. t2.micro)
         """
         for node_type in self.vendor['node-types']:
-            if self.vendor['node-types'][node_type]['name'] == instance.instance_type:
+            vmtype = node.extra.get('properties')['hardwareProfile']['vmSize']
+            if self.vendor['node-types'][node_type]['name'] == vmtype:
                 instance_type = self.vendor['node-types'][node_type]['name']
                 unit_price = self.vendor['node-types'][node_type]['price']
                 return unit_price
-            
+
     def get_host_ip(self, node_name):
         pass
+
+    def get_instance_name(self, node):
+        """Member function: get_instance_name
+        Get the name information from the instance with given ID.
+        Note: AWS doesn't use unique name for instances, instead, name is an
+        attribute stored in the tags.
+        
+         - node: a node object
+        """
+        
+        return node.name
+
+    def get_running_cost(self, verbose=True):
+
+        current_time = datetime.now(timezone.utc)
+
+        nodes = []
+        for node in self.driver.list_nodes():
+            if self.get_instance_name(node) in self.account['protected_nodes']:
+                continue
+            if node.state == "running":
+                # Get the creation time of the instance
+                creation_time_str = node.extra.get('properties')['timeCreated']  # Azure
+                if creation_time_str:
+                    # Convert the creation time from string to datetime object
+                    # Azure returns 7-digit after '.' for seconds, so need to truncate the last digit 
+                    # to cast into %Y-%m-%dT%H:%M:%S.%f%z format
+                    idx = creation_time_str.find('+')
+                    creation_time_str = creation_time_str[:idx-1] + creation_time_str[idx:]
+
+                    creation_time = datetime.strptime(creation_time_str, '%Y-%m-%dT%H:%M:%S.%f%z')
+                    
+                    # Calculate the running time
+                    running_time = current_time - creation_time
+
+                    # get the node type
+                    node_type = node.extra.get('properties')['hardwareProfile']['vmSize']
+
+                    instance_unit_cost = self.get_unit_price(node)
+                    running_cost = running_time.seconds/3600.0 * instance_unit_cost
+
+                    nodes.append([node.name, node_type, running_time, running_cost])
+
+        if verbose == True:
+            print(tabulate(nodes, headers=['Name', 'Type', 'Running Time', 'Running Cost']))
+            print("")
+
+
